Use floats for display frequency and fov

+++ common/global_module_defs.hpp
@@ -16,16 +16,16 @@ static constexpr unsigned display_width = 2560;
 static constexpr unsigned display_height = 1440;
 
 // Display refresh rate in Hz
-static constexpr unsigned display_frequency = 120;
+static constexpr float display_frequency = 120.0f;
 
 // Display period in nanoseconds
 static constexpr duration display_period = freq2period(display_frequency);
 
 // Display horizontal field-of-view in degrees
-static constexpr float display_fov_x = 45.0;
+static constexpr float display_fov_x = 45.0f;
 
 // Display vertical field-of-view in degrees
-static constexpr float display_fov_y = 45.0;
+static constexpr float display_fov_y = 45.0f;
 
 // Inter-pupilary distance (ipd) in meters
 static constexpr float ipd = 0.064f;
